[refactor] Rank enum이 순위 연산을 맡도록 수정
