+++ src/main/java/lotto/domain/ResultCalculator.java
@@ -1,21 +1,9 @@
 package lotto.domain;
 
-import java.util.Arrays;
-
 public class ResultCalculator {
     private static final long DECIMAL_TO_PERCENT_RATE = 100;
     private static final int ORIGINAL_PERCENT = 100;
     private static final int LOTTO_PRICE = 1000;
-
-    public static void findRank(int rightNumber, boolean isCorrectBonusNumber) {
-        Arrays.stream(Rank.values())
-                .filter(rank -> isSameRank(rank, rightNumber, isCorrectBonusNumber))
-                .forEach(Rank::countUp);
-    }
-
-    private static boolean isSameRank(Rank rank, int correctLottoNumber, boolean isCorrectBonusNumber) {
-        return rank.correctLottoNumber == correctLottoNumber && rank.isCorrectBonusNumber == isCorrectBonusNumber;
-    }
 
     public static long calculateEarningRate(PurchaseAmount purchaseAmount) {
         int purchaseMoney = purchaseAmount.getPurchaseNumber() * LOTTO_PRICE;
